api/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
snowflake-connector-python
google-cloud-bigquery
pandas
lightgbm
mlflow
shap
scikit-learn
numpy
matplotlib
chromadb
sentence-transformers
langchain
langchain-groq
langchain-community
groq
